Add basic makefile system

+++ src/sort/sort.c
@@ -8,7 +8,7 @@ void bubbleSort(int array[], const size_t len) {
     bool hasChanges = true;
     while (hasChanges) {
         hasChanges = false;
-        for (int i = 0; i < len - 1; ++i) {
+        for (size_t i = 0; i < len - 1; ++i) {
             if (array[i] > array[i + 1]) {
                 swap(&array[i], &array[i+1]);
                 hasChanges = true;
@@ -118,7 +118,7 @@ void quickSort(int array[], const size_t len) {
     }
 
     // Choose a random pivot
-    size_t pivot = rand() % len;
+    size_t pivot = (size_t)rand() % len;
 
     // Partition
     swap(&array[pivot], &array[len - 1]);
